requirements.txt
### dependency
tqdm
mlflow
pandas
numpy
PyYAML
Pillow
imghdr
tensorflow

### local packages -
-e . # << uncomment this to install from setup.py
